requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
alembic==1.13.1
psycopg2==2.9.9
python-dotenv==1.0.1
black==24.2.0